Lists: * Move Long List tests to Lists module
